# doc/api.yaml
openapi: 3.0.0
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
servers:
  - url: /api
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /login:
    post:
      tags:
        - Login
      security: []
      summary: Login
      description: Logins a user and returns a JWT-token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                login:
                  type: string
                  description: Username
                password:
                  type: string
                  description: Password
              required:
                - user
                - login
      responses:
        200:
          description: Successful login.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                description: JWT Token
        403:
          description: Incorrect login or password
  /signup:
    post:
      tags:
        - Signup
      security: []
      summary: Signup
      description: Signup a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                login:
                  type: string
                  minLength: 3
                  maxLength: 255
                  description: Username
                password:
                  type: string
                  format: password
                  pattern: ^[a-zA-Z0-9]{3,30}
                  description: Password
              required:
                - login
                - password
      responses:
        204:
          description: Successful signup
        400:
          description: Bad request
        409:
          description: Conflict. Login already exists
...